+++ sdks/typescript/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] - 2025-08-12
+
+### Added
+
+- getSummaries API returning structured summaries
+
+### Changed
+
+- Messages can take an optional `created_at` value, defaulting to the current
+  time (UTC ISO 8601)
+
 ## [1.3.0] - 2025-08-04
 
 ### Added